CN-B-9: change the CountDTO int type to Long

+++ src/main/java/ru/sirius/concordia/chat/repository/ChatMessageRepository.java
@@ -8,7 +8,7 @@ import java.util.List;
 
 public interface ChatMessageRepository extends MongoRepository<ChatMessage, String> {
 
-    int countBySenderIdAndRecipientIdAndStatus(
+    Long countBySenderIdAndRecipientIdAndStatus(
             String senderId,
             String recipientId,
             MessageStatus status
